clean up code pt 2
