+++ src/plugins/explore/public/components/tabs/action_bar/patterns_settings/patterns_settings_popover_button.tsx
@@ -19,7 +19,6 @@ export const PatternsSettingsPopoverButton = () => {
         <EuiButtonIcon
           size="s"
           data-test-subj="patternsSettingButton"
-          // disabled={isLoading}
           iconType="gear"
           onClick={openPopover}
         />
